refactor mille input for neuland
